tests/wpt/web-platform-tests/shadow-dom/elements-and-dom-objects/shadowroot-object/shadowroot-methods/test-002.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_10_01_02_02</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#shadow-root-methods">
<meta name="assert" content="ShadowRoot Object: NodeList getElementsByClassName(DOMString className) method">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
test(function () {

	var d = newHTMLDocument();

    var el = d.createElement('div');
    d.body.appendChild(el);

    var s = el.createShadowRoot();

    assert_equals(s.getElementsByClassName('clazz').length, 0, 'ShadowRoot getElementsByClassName() ' +
    		'method should return empty list if there\'s no matching child elements');

}, 'A_10_01_02_02_T01');



test(function () {

	var d = newHTMLDocument();

    var el = d.createElement('div');
    d.body.appendChild(el);

    var s = el.createShadowRoot();

    var child = d.createElement('span');
    child.setAttribute('class', 'clazz');
    s.appendChild(child);

    assert_equals(s.getElementsByClassName('clazz').length, 1, 'ShadowRoot getElementsByClassName() ' +
    		'method should return matching child element');

}, 'A_10_01_02_02_T02');


test(function () {

	var d = newHTMLDocument();

    var el = d.createElement('div');
    d.body.appendChild(el);

    var s = el.createShadowRoot();

    var child = d.createElement('span');
    child.setAttribute('class', 'clazz');
    s.appendChild(child);

    var child2 = d.createElement('span');
    child2.setAttribute('class', 'clazz');
    s.appendChild(child2);

    assert_equals(s.getElementsByClassName('clazz').length, 2, 'ShadowRoot getElementsByClassName() ' +
    		'method should return matching child elements');

}, 'A_10_01_02_02_T03');
</script>
</body>
</html>
